feat(investigation table) - moved complexity & tranfer icon indicatores to separate column
