+++ client/src/components/App/Content/LandingPage/InvestigationTable/InvestigationIndicatorsColumn/InvestigationIndicatorsColumnStyles.ts
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/styles';
+
+const useStyles = makeStyles(() => ({
+    columnContainer: {
+        display: 'flex',
+        alignItems: 'center'
+    },
+}));
+
+export default useStyles;
