+++ client/src/components/App/Content/LandingPage/InvestigationTable/useInvestigationTable.ts
@@ -491,6 +491,7 @@ const useInvestigationTable = (parameters: useInvestigationTableParameters): use
     const convertToIndexedRow = (row: InvestigationTableRow): IndexedInvestigationData => {
         return {
             [TableHeadersNames.color]: '',
+            [TableHeadersNames.rowIndicators]: '',
             [TableHeadersNames.multipleCheck]: row.isChecked,
             [TableHeadersNames.epidemiologyNumber]: row.epidemiologyNumber,
             [TableHeadersNames.coronaTestDate]: getFormattedDate(row.coronaTestDate),
